requirements.txt
python-binance
sqlalchemy